refactor:  added multi module handling
