+++ app/build.gradle
@@ -64,6 +64,10 @@ android
             dependencies {
                 // Questa parte racchiude la collezione di librerie che useremo
                 implementation fileTree(dir: "libs", include: ["*.jar"])
+                //
+                implementation project(":di")
+                //
+                implementation project(":domain")
                 // Ci consente di mettere il jar della libreria direttamente nella cartella libs
                 // Le librerie con implementations vengono scaricate da gradle
                 implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
@@ -90,13 +94,8 @@ android
                 implementation "org.koin:koin-androidx-fragment:$koin_version"
                 // Kotlin coroutines
                 implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.2'
-                // Retrofit per gestire il Json
-                implementation 'com.squareup.retrofit2:retrofit:2.9.0'
                 // Gson: Parsing per Json
-                implementation 'com.google.code.gson:gson:2.8.6'
-                implementation 'com.squareup.retrofit2:converter-gson:2.5.0'
-                // Vediamo il log nelle chiamate di rete
-                implementation 'com.squareup.okhttp3:logging-interceptor:3.12.0'
+                implementation "com.google.code.gson:gson:$gson_version"
                 // Timber
                 implementation 'com.jakewharton.timber:timber:4.7.1'
                 implementation 'androidx.legacy:legacy-support-v4:1.0.0'
