+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teamc/MyApplication.kt
@@ -2,6 +2,7 @@ package com.ivanmorgillo.corsoandroid.teamc
 
 import android.app.Application
 import android.os.StrictMode
+import com.blps.aagj.cookbook.di.networkingKoinModule
 import com.ivanmorgillo.corsoandroid.teamc.firebase.CrashReportingTree
 import com.ivanmorgillo.corsoandroid.teamc.firebase.LineNumberDebugTree
 import org.koin.android.ext.koin.androidContext
@@ -18,7 +19,7 @@ class MyApplication : Application() {
         startKoin {
             androidLogger()
             androidContext(this@MyApplication)
-            modules(appModule)
+            modules(appModule, networkingKoinModule)
         }
         /**
          * nell'app dell'utente si crasha lui non avra niente
