+++ build.gradle
@@ -4,6 +4,7 @@ buildscript {
         koin_version = '2.2.2'
         lifecycle_version = "2.3.0"
         nav_version = "2.3.3"
+        gson_version = "2.8.6"
     }
 
     repositories {
