+++ di/src/main/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<manifest package="com.blps.aagj.cookbook.di">
+
+    /
+</manifest>
