+++ domain/src/main/java/com/blps/aagj/cookbook/domain/RecipeAPI.kt
@@ -1,0 +1,12 @@
+package com.blps.aagj.cookbook.domain
+
+import com.ivanmorgillo.corsoandroid.teamc.home.AllRecipesByAreaResult
+import com.ivanmorgillo.corsoandroid.teamc.network.LoadRecipesResult
+
+interface RecipeAPI {
+    @Suppress("TooGenericExceptionCaught")
+    suspend fun loadRecipes(area: String): LoadRecipesResult
+
+    @Suppress("TooGenericExceptionCaught")
+    suspend fun loadAllRecipesByArea(): AllRecipesByAreaResult
+}
