+++ domain/src/main/java/com/blps/aagj/cookbook/domain/RecipeDetail.kt
