+++ networking/src/main/java/com/blps/aagj/cookbook/networking/RecipeService.kt
@@ -1,8 +1,5 @@
-package com.ivanmorgillo.corsoandroid.teamc.network
+package com.blps.aagj.cookbook.networking
 
-import com.ivanmorgillo.corsoandroid.teamc.network.detail.RecipeDetailDTO
-import com.ivanmorgillo.corsoandroid.teamc.network.home.AreaDTO
-import com.ivanmorgillo.corsoandroid.teamc.network.home.RecipeDTO
 import retrofit2.http.GET
 import retrofit2.http.Query
